+++ answer/06_第6回目/ex02/ex02_test.go
@@ -1,12 +1,36 @@
 package main
 
-import "testing"
+import (
+	"testing"
+)
 
-func TestDuck(t *testing.T) {
-	duck := &Duck{"tarou"}
-	actual := duck.name
-	expected := "tarou"
-	if actual != expected {
-		t.Errorf("got: %v\nwant: %v", actual, expected)
+// 参照：https://qiita.com/high5/items/4e2580241039c950e1c4
+func calc(a, b int) (int, int, int, float32) {
+	plus := a + b
+	minus := a - b
+	multi := a * b
+	divi := float32(a) / float32(b)
+	return plus, minus, multi, divi
+}
+
+func Testcalc(t *testing.T) {
+
+	a, b := 1, 2
+
+	Plus_want := 3
+	Minus_want := -1
+	Multi_want := 2
+	Divi_want := float32(0.5)
+
+	plus, minus, multi, divi := calc(a, b)
+
+	if plus != Plus_want {
+		t.Errorf("%d is not Plus ans", plus)
+	} else if minus != Minus_want {
+		t.Errorf("%d is not Minus ans", minus)
+	} else if multi != Multi_want {
+		t.Errorf("%d is not Multi ans", multi)
+	} else if divi != Divi_want {
+		t.Errorf("%d is not Divi ans", divi)
 	}
 }
